Fixed deleting a character not deleting the char properly from the owning account's list. Fixed UIType.getVal() overflowing. Added a small part of the oz shop (just to test items that need points).

+++ src/main/java/net/swordie/ms/enums/UIType.java
@@ -325,13 +325,13 @@ public enum UIType {
     UI_NOT_DEFINED(-1),
     ;
     
-    private byte val;
+    private int val;
     
     UIType(int val) {
-        this.val = (byte) val;
+        this.val =  val;
     }
 
-    public byte getVal() {
+    public int getVal() {
         return val;
     }
 }
